Create cola
This is a program that functions like a soda machines. The user chooses a drink from the menu by entering the number that represents the beverage they desire. The machine then takes the payment and returns the change.

If the user inputs a choice that is not on the menu, the machine will refund them.