Create guess-the-number
This program generates a random number between 1-100. It asks the user to guess that number. As the user guesses, the program gives clues to tell the user if the guess is too high or too low, while continuing to loop. If the user correctly guesses the number, the program alerts them of that and exits.